Change in throttle valve surge tank average velocity
The friction factor due to diffusion in dissimilarity in the geometry from the throat diameter to the main surge tank diameter is improved using the average velocity depending on the dynamic average area as the flow passes from the throat towards the main diameter region. Furthermore, the discontinuity in the velocity when simulating the model for water level below the throat has been remove making the average velocity of water inside the tank continuous at the junction of throat and the main diameter area continuous by addition of the dynamic area. Still when L_t<<L and D_t<<D the simulation is incorrect when the water level is below the throat however, when L_t tend to D and D_t tends to D the simulation corresponds to more accurate. This confirmation was done based on the simple surge tank model. The model validation for this kind of surge tank having dissimilar geometry in the entry is often validated with simulation results from the simple surge tank.